extracts types into own file
